Extract Places class
Extracts the functionality from CountryDivisions so it can be reused by
other classes that are querying for similar data.
